requirements.txt
google-api-python-client==2.108.0
openai==1.6.1
pyairtable==2.2.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-mock==3.12.0
responses==0.24.1
